proyectos/proyecto-5-artwork-verifier/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pdf2image==1.17.0
Pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
scikit-image==0.22.0
numpy==1.26.2
python-dotenv==1.0.0
